backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Utilities
python-multipart>=0.0.6
email-validator>=2.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
